+++ lib/netzke/basepack/grid/services.rb
@@ -8,7 +8,7 @@ module Netzke
           {}.tap do |res|
             records = get_records(params)
             res[:data] = records.map{|r| data_adapter.record_to_array(r, final_columns(:with_meta => true))}
-            res[:total] = count_records(params)  if config[:enable_pagination]
+            res[:total] = count_records(params)
           end
         end
 
@@ -72,11 +72,6 @@ module Netzke
         def get_records(params)
           params[:filters] = normalize_filters(params[:filters]) if params[:filters]
           params[:query] = normalize_query(params[:query]) if params[:query].present?
-          if config[:enable_pagination]
-            params[:limit] = config[:rows_per_page]
-          else
-            params.delete(:limit)
-          end
           params[:scope] = config[:scope] # note, params[:scope] becomes ActiveSupport::HashWithIndifferentAccess
 
           data_adapter.get_records(params, final_columns)