+++ spec/rails_app/app/components/grid/associations.rb
@@ -2,6 +2,7 @@ class Grid::Associations < Netzke::Basepack::Grid
   def configure(c)
     super
     c.model = 'Book'
+    c.edit_inline = true
     c.columns = [:title, :author__first_name, :author__prize_count]
   end
 end