+++ spec/rails_app/app/components/grid/buffered.rb
@@ -1,0 +1,7 @@
+class Grid::Buffered < Netzke::Basepack::Grid
+  def configure(c)
+    super
+    c.model = 'Author'
+    c.columns = [:first_name]
+  end
+end