+++ spec/rails_app/app/components/grid/crud_paging.rb
@@ -1,8 +1,9 @@
-class Grid::Crud < Netzke::Basepack::Grid
+class Grid::CrudPaging < Netzke::Basepack::Grid
   def configure(c)
     super
     c.model = 'Book'
     c.columns = [:author__name, :title] # do not modify
+    c.paging = true
     c.persistence = true
     c.store_config = {:sorters => {property: :id}}
   end