+++ spec/rails_app/app/components/grid/multisorting.rb
@@ -3,6 +3,7 @@ class Grid::Multisorting < Netzke::Basepack::Grid
     super
     c.model = 'Book'
     c.columns = [:author__last_name, :title, :exemplars]
-    c.data_store.sorters = [:exemplars, :title, :author__last_name]
+    # c.paging = true
+    c.store_config = {sorters: [:exemplars, :title, :author__last_name]}
   end
 end