+++ spec/rails_app/app/components/grid/sorting.rb
@@ -3,6 +3,6 @@ class Grid::Sorting < Netzke::Basepack::Grid
     super
     c.model = 'Book'
     c.columns = [:author__last_name, :title, :exemplars]
-    c.data_store.sorters = {property: :exemplars, direction: 'DESC'}
+    c.store_config = {sorters: {property: :exemplars, direction: 'DESC'}}
   end
 end