+++ spec/rails_app/app/example_components/author_grid.rb
@@ -2,6 +2,5 @@ class AuthorGrid < Netzke::Basepack::Grid
   def configure(c)
     super
     c.model = "Author"
-    c.scope = ->(r) { r.order(:first_name) }
   end
 end