+++ spec/rails_app/app/example_components/book_grid.rb
@@ -6,28 +6,19 @@ class BookGrid < Netzke::Basepack::Grid
     super
   end
 
-  column :author__name do |c|
-    c.sorting_scope = :sorted_by_author_name
-    c.editor = {min_chars: 1}
-  end
-
-  client_class do |c|
-    c.init_component = <<-JS
-      function() {
-        this.callParent();
-        this.getStore().on("remove", function(){ console.log('remove'); });
-      }
-    JS
-  end
+  # column :author__name do |c|
+  #   c.sorting_scope = :sorted_by_author_name
+  #   c.editor = {min_chars: 1}
+  # end
 
   # crafting a static combobox column; TODO: include a prebuilt one in Basepack
-  column :rating do |c|
-    c.editor = {
-      :trigger_action => :all,
-      :xtype => :combo,
-      :store => [[0, "---"], [1, "Good"], [2, "Average"], [3, "Poor"]]
-    }
+  # column :rating do |c|
+  #   c.editor = {
+  #     :trigger_action => :all,
+  #     :xtype => :combo,
+  #     :store => [[0, "---"], [1, "Good"], [2, "Average"], [3, "Poor"]]
+  #   }
 
-    c.renderer = "function(v){return ['', 'Good', 'Average', 'Poor'][v];}"
-  end
+  #   c.renderer = "function(v){return ['', 'Good', 'Average', 'Poor'][v];}"
+  # end
 end