+++ spec/rails_app/app/example_components/book_grid_with_paging.rb
@@ -2,7 +2,6 @@ class BookGridWithPaging < Netzke::Basepack::Grid
   def configure(c)
     c.model = "Book"
     c.title = "Books with paging"
-    c.rows_per_page = 2
     super
   end
 end